pyproject.toml
[project]
name = "open-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "gradium>=0.5.4",
    "langchain-huggingface>=1.2.0",
    "langchain-nvidia-ai-endpoints>=1.0.2",
    "langgraph>=1.0.7",
    "numpy>=2.4.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "streamlit>=1.53.1",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
]
